Server/requirements.txt
numpy==1.12.1
tensorflow==1.0.0
keras==2.1.4 
flask==1.0.2
pillow==5.1.0
h5py==2.8.0
matplotlib==2.2.4